+++ docs/plugins.md
@@ -2,16 +2,16 @@
 
 Psalm can be extended through plugins to find domain-specific issues.
 
-Plugins may implement one of (or more than one of) `Psalm\PluginApi\Hook\*` interface(s).
+Plugins may implement one of (or more than one of) `Psalm\Plugin\Hook\*` interface(s).
 
 ```php
 <?php
-class SomePlugin implements \Psalm\PluginApi\Hook\AfterStatementAnalysisInterface
+class SomePlugin implements \Psalm\Plugin\Hook\AfterStatementAnalysisInterface
 {
 }
 ```
 
-`Psalm\PluginApi\Hook\*` offers six interfaces that you can implement:
+`Psalm\Plugin\Hook\*` offers six interfaces that you can implement:
 
  - `AfterStatementAnalysisInterface` - called after Psalm evaluates each statement
  - `AfterExpressionAnalysisInterface` - called after Psalm evaluates each expression
@@ -56,7 +56,7 @@ Composer-based plugin is a composer package which conforms to these requirements
 
 1. Its `type` field is set to `psalm-plugin`
 2. It has `extra.psalm.pluginClass` subkey in its `composer.json` that reference an entry-point class that will be invoked to register the plugin into Psalm runtime.
-3. Entry-point class implements `Psalm\PluginApi\PluginEntryPointInterface`
+3. Entry-point class implements `Psalm\Plugin\PluginEntryPointInterface`
 
 ### Using skeleton project
 
@@ -64,10 +64,10 @@ Run `composer create-project weirdan/psalm-plugin-skeleton:dev-master your-plugi
 
 ### Upgrading file-based plugin to composer-based version
 
-Create new plugin project using skeleton, then pass the class name of you file-based plugin to `registerHooksFromClass()` method of the `Psalm\PluginApi\RegistrationInterface` instance that was passed into your plugin entry point's `__invoke()` method. See the [conversion example](https://github.com/vimeo/psalm/examples/plugins/composer-based/echo-checker/).
+Create new plugin project using skeleton, then pass the class name of you file-based plugin to `registerHooksFromClass()` method of the `Psalm\Plugin\RegistrationInterface` instance that was passed into your plugin entry point's `__invoke()` method. See the [conversion example](https://github.com/vimeo/psalm/examples/plugins/composer-based/echo-checker/).
 
 ### Registering stub files
 
-Use `Psalm\PluginApi\RegistrationInterface::addStubFile()`. See the [sample plugin] (https://github.com/weirdan/psalm-doctrine-collections/).
+Use `Psalm\Plugin\RegistrationInterface::addStubFile()`. See the [sample plugin] (https://github.com/weirdan/psalm-doctrine-collections/).
 
 Stub files provide a way to override third-party type information when you cannot add Psalm's extended docblocks to the upstream source files directly.
