+++ examples/plugins/ClassUnqualifier.php
@@ -5,7 +5,7 @@ use Psalm\Codebase;
 use Psalm\CodeLocation;
 use Psalm\FileManipulation\FileManipulation;
 use Psalm\FileSource;
-use Psalm\PluginApi\Hook\AfterClassLikeExistenceCheckInterface;
+use Psalm\Plugin\Hook\AfterClassLikeExistenceCheckInterface;
 use Psalm\StatementsSource;
 use Psalm\Type;
