+++ examples/plugins/composer-based/echo-checker/PluginEntryPoint.php
@@ -1,13 +1,13 @@
 <?php
 namespace Psalm\Example\Plugin\ComposerBased;
 
-use Psalm\PluginApi;
+use Psalm\Plugin;
 use SimpleXMLElement;
 
-class PluginEntryPoint implements PluginApi\PluginEntryPointInterface
+class PluginEntryPoint implements Plugin\PluginEntryPointInterface
 {
     /** @return void */
-    public function __invoke(PluginApi\RegistrationInterface $registration, ?SimpleXMLElement $config = null)
+    public function __invoke(Plugin\RegistrationInterface $registration, ?SimpleXMLElement $config = null)
     {
         require_once __DIR__ . '/EchoChecker.php';
         $registration->registerHooksFromClass(EchoChecker::class);
