+++ src/Psalm/Plugin/Hook/AfterClassLikeExistenceCheckInterface.php
@@ -1,22 +1,21 @@
 <?php
-namespace Psalm\PluginApi\Hook;
+namespace Psalm\Plugin\Hook;
 
-use PhpParser\Node\Stmt\ClassLike;
+use Psalm\CodeLocation;
 use Psalm\Codebase;
 use Psalm\FileManipulation\FileManipulation;
 use Psalm\StatementsSource;
-use Psalm\Storage\ClassLikeStorage;
 
-interface AfterClassLikeVisitInterface
+interface AfterClassLikeExistenceCheckInterface
 {
     /**
      * @param  FileManipulation[] $file_replacements
      *
      * @return void
      */
-    public static function afterClassLikeVisit(
-        ClassLike $stmt,
-        ClassLikeStorage $storage,
+    public static function afterClassLikeExistenceCheck(
+        string $fq_class_name,
+        CodeLocation $code_location,
         StatementsSource $statements_source,
         Codebase $codebase,
         array &$file_replacements = []
