+++ src/Psalm/Plugin/Hook/AfterMethodCallAnalysisInterface.php
@@ -1,5 +1,5 @@
 <?php
-namespace Psalm\PluginApi\Hook;
+namespace Psalm\Plugin\Hook;
 
 use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\MethodCall;
