+++ src/Psalm/Plugin/PluginEntryPointInterface.php
@@ -1,7 +1,8 @@
 <?php
-namespace Psalm\PluginApi;
+namespace Psalm\Plugin;
 
 use SimpleXMLElement;
+use Psalm\Plugin\RegistrationInterface;
 
 interface PluginEntryPointInterface
 {
